Split Slack API into two files
